Msg::Func: daemon fork is moved after Exe initialize;

+++ script/libmsgfunc.rb
@@ -57,30 +57,37 @@ module CIAX
     # Switch error output to file
     def err2file(tag)
       return if $stderr.is_a? Tee
-      fname = vardir('log') + 'error_' + tag + today + '.out'
-      $stderr = Tee.new(fname)
-      Process.daemon(true, true)
+      outfile = vardir('log') + 'error_' + tag + today + '.out'
+      $stderr = Tee.new(outfile)
     end
 
-    # Reloadable by HUP signal
-    def daemon(tag)
-      # Set ARGS in opt file
-      dir = vardir('run')
-      optfile = "#{dir}#{tag}.opt"
-      pidfile = "#{dir}#{tag}.pid"
-      IO.foreach(pidfile) do |line|
+    def new_pid(base)
+      Process.daemon(true, true)
+      IO.write(base + '.pid', $PROCESS_ID)
+    end
+
+    def kill_pid(base)
+      return unless test('r', base + '.pid')
+      IO.foreach(base + '.pid') do |line|
         pid = line.to_i
         next unless pid > 0
         begin
           Process.kill(:TERM, pid)
         rescue
+          nil
         end
-      end if test('r', pidfile)
-      IO.write(pidfile, $PROCESS_ID)
+      end
+    end
+
+    # Reloadable by HUP signal
+    def daemon(tag)
+      # Set ARGS in opt file
+      base = vardir('run') + tag
+      kill_pid(base)
       begin
-        load optfile if test('r', optfile)
+        load optfile if test('r', base + '.opt')
         exe = yield
-        err2file(tag)
+        err2file(tag) && new_pid(base)
         exe.server
         sleep
       rescue SignalException
@@ -88,6 +95,7 @@ module CIAX
       end
     end
 
+    # append str to stderr and file like tee
     class Tee < IO
       def initialize(fname)
         super(2)
